+++ Predicate.cpp
@@ -14,9 +14,9 @@ Predicate::Predicate(std::string id, std::vector<Parameter> parameters) {
 std::string Predicate::toString() {
     std::string returnable = "";
     returnable.append(id + "(");
-    for (int i = 0; i < parameters.size(); i++) {
+    for (int i = 0; i < (int) parameters.size(); i++) {
         returnable.append(parameters[i].toString());
-        if (i < parameters.size() - 1) {
+        if (i < (int) parameters.size() - 1) {
             returnable.append(",");
         } else {
             returnable.append(")");
