+++ Rule.cpp
@@ -15,9 +15,9 @@ Rule::Rule(Predicate headPredicate, std::vector<Predicate> predicates) {
 std::string Rule::toString() {
     std::string returnable = "";
     returnable.append(headPredicate.toString() + " :- ");
-    for (int i = 0; i < predicates.size(); i++) {
+    for (int i = 0; i < (int)predicates.size(); i++) {
         returnable.append(predicates[i].toString());
-        if (i < predicates.size() - 1) {
+        if (i < (int) (predicates.size() - 1)) {
             returnable.append(",");
         } else {
             returnable.append(".");
